api/requirements.txt
flask>=2.0.0       
keras==2.6.0
numpy==1.22.3
pillow==9.0.1
requests==2.27.1